task 1

+++ assignment_improvement_exercise edit.py
@@ -4,7 +4,7 @@
 
 import math
 
-radius = int(input("please enter the radius of the circle: "))
+radius = float(input("please enter the radius of the circle: "))
 
 circumf = 2* math.pi * radius
 circumf = round(circumf,2)
